samples: updates face and people to v1
