+++ samples/snippets/src/main/java/video/DetectFacesGcs.java
@@ -14,45 +14,37 @@
  * limitations under the License.
  */
 
-package beta.video;
+package video;
 
-// [START video_detect_faces_beta]
+// [START video_detect_faces_gcs]
 
 import com.google.api.gax.longrunning.OperationFuture;
-import com.google.cloud.videointelligence.v1p3beta1.AnnotateVideoProgress;
-import com.google.cloud.videointelligence.v1p3beta1.AnnotateVideoRequest;
-import com.google.cloud.videointelligence.v1p3beta1.AnnotateVideoResponse;
-import com.google.cloud.videointelligence.v1p3beta1.DetectedAttribute;
-import com.google.cloud.videointelligence.v1p3beta1.FaceDetectionAnnotation;
-import com.google.cloud.videointelligence.v1p3beta1.FaceDetectionConfig;
-import com.google.cloud.videointelligence.v1p3beta1.Feature;
-import com.google.cloud.videointelligence.v1p3beta1.TimestampedObject;
-import com.google.cloud.videointelligence.v1p3beta1.Track;
-import com.google.cloud.videointelligence.v1p3beta1.VideoAnnotationResults;
-import com.google.cloud.videointelligence.v1p3beta1.VideoContext;
-import com.google.cloud.videointelligence.v1p3beta1.VideoIntelligenceServiceClient;
-import com.google.cloud.videointelligence.v1p3beta1.VideoSegment;
-import com.google.protobuf.ByteString;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import com.google.cloud.videointelligence.v1.AnnotateVideoProgress;
+import com.google.cloud.videointelligence.v1.AnnotateVideoRequest;
+import com.google.cloud.videointelligence.v1.AnnotateVideoResponse;
+import com.google.cloud.videointelligence.v1.DetectedAttribute;
+import com.google.cloud.videointelligence.v1.FaceDetectionAnnotation;
+import com.google.cloud.videointelligence.v1.FaceDetectionConfig;
+import com.google.cloud.videointelligence.v1.Feature;
+import com.google.cloud.videointelligence.v1.TimestampedObject;
+import com.google.cloud.videointelligence.v1.Track;
+import com.google.cloud.videointelligence.v1.VideoAnnotationResults;
+import com.google.cloud.videointelligence.v1.VideoContext;
+import com.google.cloud.videointelligence.v1.VideoIntelligenceServiceClient;
+import com.google.cloud.videointelligence.v1.VideoSegment;
 
-public class DetectFaces {
+public class DetectFacesGcs {
 
-  public static void detectFaces() throws Exception {
+  public static void detectFacesGcs() throws Exception {
     // TODO(developer): Replace these variables before running the sample.
-    String localFilePath = "resources/googlework_short.mp4";
-    detectFaces(localFilePath);
+    String gcsUri = "gs://cloud-samples-data/video/googlework_short.mp4";
+    detectFacesGcs(gcsUri);
   }
 
-  // Detects faces in a video stored in a local file using the Cloud Video Intelligence API.
-  public static void detectFaces(String localFilePath) throws Exception {
+  // Detects faces in a video stored in Google Cloud Storage using the Cloud Video Intelligence API.
+  public static void detectFacesGcs(String gcsUri) throws Exception {
     try (VideoIntelligenceServiceClient videoIntelligenceServiceClient =
         VideoIntelligenceServiceClient.create()) {
-      // Reads a local video file and converts it to base64.
-      Path path = Paths.get(localFilePath);
-      byte[] data = Files.readAllBytes(path);
-      ByteString inputContent = ByteString.copyFrom(data);
 
       FaceDetectionConfig faceDetectionConfig =
           FaceDetectionConfig.newBuilder()
@@ -65,7 +57,7 @@ public class DetectFaces {
 
       AnnotateVideoRequest request =
           AnnotateVideoRequest.newBuilder()
-              .setInputContent(inputContent)
+              .setInputUri(gcsUri)
               .addFeatures(Feature.FACE_DETECTION)
               .setVideoContext(videoContext)
               .build();
@@ -80,7 +72,7 @@ public class DetectFaces {
       // Gets annotations for video
       VideoAnnotationResults annotationResult = response.getAnnotationResultsList().get(0);
 
-      // Annotations for list of faces detected, tracked and recognized in video.
+      // Annotations for list of people detected, tracked and recognized in video.
       for (FaceDetectionAnnotation faceDetectionAnnotation :
           annotationResult.getFaceDetectionAnnotationsList()) {
         System.out.print("Face detected:\n");
@@ -109,4 +101,4 @@ public class DetectFaces {
     }
   }
 }
-// [END video_detect_faces_beta]
+// [END video_detect_faces_gcs]
