+++ samples/snippets/src/test/java/video/DetectFacesIT.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package beta.video;
+package video;
 
 import static com.google.common.truth.Truth.assertThat;
 
@@ -23,9 +23,11 @@ import java.io.PrintStream;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import video.DetectFaces;
 
 @SuppressWarnings("checkstyle:AbbreviationAsWordInName")
-public class DetectFacesGcsIT {
+public class DetectFacesIT {
+
   private ByteArrayOutputStream bout;
   private PrintStream out;
   private PrintStream originalPrintStream;
@@ -46,8 +48,8 @@ public class DetectFacesGcsIT {
   }
 
   @Test
-  public void testDetectFacesGcs() throws Exception {
-    DetectFacesGcs.detectFacesGcs("gs://cloud-samples-data/video/googlework_short.mp4");
+  public void testDetectFaces() throws Exception {
+    DetectFaces.detectFaces("resources/googlework_short.mp4");
     String got = bout.toString();
     assertThat(got).contains("Face detected:");
   }
